moved Rental buttons/banner from App.js into its own component
